Testing including external jars in language definition project
First do these steps
https://xn--lubisoft-0za.gmbh/en/articles/using-maven-artifacts-in-pde-rcp-and-tycho-builds/

Then update the pom with the dependency and the Eclipse classpath

Then the import should work in a class

Need to set default target platform from target file, then MANIFEST can
include the new library turned into a plug-in

One reason for "No application id found" is that the id for launching
the IDE product is: "org.eclipse.platform.ide"

Everything except the standalone test passes.
The jar isn't included in the plug-in and so the install doesn't go
through.