Removed dependencies on the jar wrapped by a plug-in
I'm doing this because I can't find information on how to import the
exported plug-in. All the steps I find through Google are 10+ years old.
The general idea is to use the install new software wizard which I do
for the plug-in itself but it doesn't work for the jar wrapped plug-in.
The latter seems to be missing some information and the wizard outputs
some error.