Added cucumber and surefire plug-ins to tests project
I tried to integrate the Xtext unit testing with Cucumber but there's
some weird assertion error. For now I assume it's because there's two
types of tests.

I think I'll have to remove the Cucumber stuff because it's not being
invoked by Maven. The tycho stuff seems to do its own thing, ie not
recognise the Maven dependencies.

I can try and generate the Xtext type test automation like I would for
PST/PSM stuff. I'll have to move the Cucumber specs to their own project
and use a Maven goal to do the transform and supply different
transformers like at work. That's a nice to have exercise for another
time.

The other option is to directly invoke the surefire goal instead of the
test phase which I don't like.

The third option is to move the UML and Java parser stuff out into their
own jar and have the xtext project reference that jar. But that jar
would need a reference to the xtext language definition setting up a
cyclical dependency. Will think more about this.

TODO
1. Publish plug-in
2. Create feature files for the tests
3. Make glue and objects code
4. Update language grammar and validation to continue making tests
5. Publish plug-in