Integrated mbt-transformer into Maven goals
Layer 2 objects are either EdgeTestObject or VerticeTestObject. Both
are TestObject but only EdgeTestObject should have the traverseEdgeTo
(execute) method.

If each edge is represented by one class, then there only needs to be
one execute method implemented which is the current code generated.
However if each edge is represented as a method in the vertice classes,
then just as set/assertAttributes defines the input/output in the FSM
dynamically, the execute method should select the edge dynamically.

The latter might make sense for web-pages to reuse xpath definitions in
the objects like to locate a search box to assert its there and then to
fill it in. However for web-services it's better to have two separate
classes, one for request payload (input given for an edge) and one for
the output (vertice)

When there are multiple edges to or from a vertice I'll need to
distinguish the names. For now I can just name the object "Insurer
Details from Claim Request" vs "Insurer Details from Group Request"
which will create two separate classes for the two separate edge names.

For now I'm letting there be just one TestObject with the execute
method.

TODO
- docstring layer 2 methods should have setAttributes for "is" vs "will
be"
- test Cucumber to UML in mbt transformer
- test UML to Cucumber in mbt transformer
- Publish release versions of everything. Add Maven release plug-in to
mbt-transformer, mbt-validator, mbt-maven-plugin